+++ src/main/java/codechicken/nei/RecipeSearchField.java
@@ -3,7 +3,6 @@ package codechicken.nei;
 import net.minecraft.client.Minecraft;
 import net.minecraft.util.EnumChatFormatting;
 
-import codechicken.nei.SearchField.SearchTextFormatter;
 import codechicken.nei.api.ItemFilter;
 import codechicken.nei.util.TextHistory;
 
@@ -19,7 +18,7 @@ public abstract class RecipeSearchField extends TextField {
     @Override
     protected void initInternalTextField() {
         field = new FormattedTextField(Minecraft.getMinecraft().fontRenderer, 0, 0, 0, 0);
-        ((FormattedTextField) field).setFormatter(new SearchTextFormatter());
+        ((FormattedTextField) field).setFormatter(new SearchTextFormatter(SearchField.searchParser));
         field.setMaxStringLength(maxSearchLength);
         field.setCursorPositionZero();
     }
