+++ src/main/java/codechicken/nei/SubsetWidget.java
@@ -28,8 +28,8 @@ import codechicken.lib.vec.Rectangle4i;
 import codechicken.nei.ItemList.AnyMultiItemFilter;
 import codechicken.nei.ItemList.ItemsLoadedCallback;
 import codechicken.nei.ItemList.NothingItemFilter;
-import codechicken.nei.SearchField.ISearchParserProvider;
-import codechicken.nei.SearchField.SearchMode;
+import codechicken.nei.SearchTokenParser.ISearchParserProvider;
+import codechicken.nei.SearchTokenParser.SearchMode;
 import codechicken.nei.api.API;
 import codechicken.nei.api.ItemFilter;
 import codechicken.nei.api.ItemFilter.ItemFilterProvider;
