+++ src/main/resources/assets/nei/lang/en_US.lang
@@ -210,12 +210,6 @@ nei.options.inventory.search.subsetsSearchMode=Item Subsets Search
 nei.options.inventory.search.subsetsSearchMode.0=Always
 nei.options.inventory.search.subsetsSearchMode.1=Prefix: %s
 nei.options.inventory.search.subsetsSearchMode.2=Never
-nei.options.inventory.centerSearchWidget=Center Search Widget
-nei.options.inventory.centerSearchWidget.true=Yes
-nei.options.inventory.centerSearchWidget.false=No
-nei.options.inventory.focusSearchWidgetOnOpen=Focus Search Widget on Open
-nei.options.inventory.focusSearchWidgetOnOpen.true=Yes
-nei.options.inventory.focusSearchWidgetOnOpen.false=No
 nei.options.inventory.firstInvCloseClosesInSearch=First Inventory Close Closes in Search
 nei.options.inventory.firstInvCloseClosesInSearch.true=Yes
 nei.options.inventory.firstInvCloseClosesInSearch.false=No
